feat(v0.5.0): MQ135 integration and hardware fixes
- Implement PPM persistence in database

- Fix buzzer using tone()

- Swap GRB LED pins

- Change read interval to 10s

- Update simulator and cloud sync for PPM

+++ python/web_server.py
@@ -217,7 +217,7 @@ def api_clear_hard():
 
 # ========== 供外部呼叫的函數 ==========
 
-def update_current_reading(temperature: float, humidity: float, heat_index: float = None, air_quality: int = None):
+def update_current_reading(temperature: float, humidity: float, heat_index: float = None, air_quality: float = None):
     """更新即時數據（供 main.py 呼叫）"""
     global current_reading
     current_reading = {
